+++ README.md
@@ -48,10 +48,11 @@
      - **Home Banner Slide** - [commit 04647e9](https://github.com/ld5ehom/store-ios/commit/04647e90727f0b0f90285999675564785adb7b12) : 
        - The Home storyboard view controller wasn't connected, so the storyboard was loaded to properly use the Home view controller.
        - Using a compositional layout and diffable data source to efficiently manage and display collection view items. 
-     - **Home Screen Horizontal Product Carousel** - [commit 9f83a10](https://github.com/ld5ehom/store-ios/commit/9f83a10c80d0499c1b372c3f65d21044920bffef) :  
+     - **Home Screen Recommended Products Using Horizontal Carousel** - [commit 9f83a10](https://github.com/ld5ehom/store-ios/commit/9f83a10c80d0499c1b372c3f65d21044920bffef) :  
        - Implemented a horizontal carousel component for the Home screen, where each cell displays product details such as images, titles, and prices. The carousel allows users to scroll through products horizontally. Each cell features a product image, title, and price, with the original price displayed with a strikethrough effect to highlight discounts.
-     - **Home Screen Vertical Product Carousel** :  
-       - 
+     - **Home Screen Product Display Using Vertical Scroll** : 
+       - Refactored Cell Layout Configuration: Moved the cell layout configuration from the view controller to the cell extension, resulting in cleaner code and improved modularity.
+       - Implemented Vertical Scroll Layout: Created a product list for general items using a vertical scroll layout. This arrangement allows users to scroll through the list of products efficiently, organizing them vertically for better display and navigation.
